refactor(models): isolate authored field resolution
